separated openai services into four files - utils, tools, services and functions added return types to all functions in openai services folder
